saving data functions splited
