--- opticFiberProject2021-main/OpticFiberTest_ver6/OpticFiberTest_ver1/SaveData/SaveExel.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OpticFiberTest_ver1.SaveData
-{
-    class SaveExel
-    {
-    }
-}
